Fix flags tests
