+++ pkg/flags/flags.go
@@ -35,18 +35,7 @@ import (
 	"k8s.io/ingress-nginx/internal/ingress/status"
 	ing_net "k8s.io/ingress-nginx/internal/net"
 	"k8s.io/ingress-nginx/internal/nginx"
-	"k8s.io/ingress-nginx/pkg/dataplane"
 	klog "k8s.io/klog/v2"
-)
-
-var (
-	flags = pflag.NewFlagSet("", pflag.ExitOnError)
-
-	showVersion = flags.Bool("version", false,
-		`Show release information about the NGINX Ingress controller and exit.`)
-
-	healthzPort = flags.Int("healthz-port", 10254, "Port to use for the healthz endpoint.")
-	healthzHost = flags.String("healthz-host", "", "Address to bind the healthz endpoint.")
 )
 
 // TODO: We should split the flags functions between common for all programs
@@ -56,6 +45,14 @@ var (
 func ParseFlags() (bool, *controller.Configuration, error) {
 
 	var (
+		flags = pflag.NewFlagSet("", pflag.ExitOnError)
+
+		showVersion = flags.Bool("version", false,
+			`Show release information about the NGINX Ingress controller and exit.`)
+
+		healthzPort = flags.Int("healthz-port", 10254, "Port to use for the healthz endpoint.")
+		healthzHost = flags.String("healthz-host", "", "Address to bind the healthz endpoint.")
+
 		apiserverHost = flags.String("apiserver-host", "",
 			`Address of the Kubernetes API server.
 Takes the form "protocol://address:port". If not specified, it is assumed the
@@ -392,33 +389,6 @@ https://blog.maxmind.com/2019/12/18/significant-changes-to-accessing-and-using-g
 		config.MaxmindEditionFiles = &nginx.MaxmindEditionFiles
 	}
 
-	return false, config, err
-}
-
-func ParseDataplaneFlags() (bool, *dataplane.Configuration, error) {
-
-	var (
-		grpcAddress = flags.String("grpc-host", "ingress-nginx:10000", "Address to connect to gRPC Control plane")
-	)
-
-	flags.Parse(os.Args)
-
-	// Workaround for this issue:
-	// https://github.com/kubernetes/kubernetes/issues/17162
-	flag.CommandLine.Parse([]string{})
-
-	pflag.VisitAll(func(flag *pflag.Flag) {
-		klog.V(2).InfoS("FLAG", flag.Name, flag.Value)
-	})
-
-	if *showVersion {
-		return true, nil, nil
-	}
-
-	var err error
-	config := &dataplane.Configuration{
-		GRPCAddress: *grpcAddress,
-	}
 	return false, config, err
 }
